De-Helgi-fied the Public licence
De-Helgi-fied the Public licence and Fisedush-ified the licence instead so it's like Fisedush is our company/org. I did this so there's no debate who should be named there (and other places where names have to be put). And it is more professional to put a company rather than a name of a person (or three persons)